#203599
#204116
Created unit test for DataMining library (most of the functionalities are also already covered by casa testing)
Fixed warnings and enabled -Werror with gcc for libraries that are already warning-free

Related work items: #204116, #203599

+++ development/libraries/SerialDataAccess/src/SerialGrid.C
@@ -245,10 +245,10 @@ double SerialGrid::getSurface (unsigned int i, unsigned int j) const
 {
    double surface = m_surface;
 
-   if (i == 0 || i == numI () - 1)
+   if (i == 0 || static_cast<int>(i) == numI () - 1)
       surface /= 2;
 
-   if (j == 0 || j == numJ () - 1)
+   if (j == 0 || static_cast<int>(j) == numJ () - 1)
       surface /= 2;
 
    return surface;
